safari-600.7-branch/LayoutTests/platform/mac/accessibility/text-marker-for-bounds-expected.txt

Text text text text text text text text text text.
This tests that endTextMarkerForBounds and startTextMarkerForBounds work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rangeFromBounds is rangeFromElement
PASS successfullyParsed is true

TEST COMPLETE
